user_management/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyMySQL==1.1.0
Pillow==9.5.0
PyJWT==2.6.0
cryptography==41.0.7 